requirements.prod.txt
# ── Core Flask ───────────────────────────────────────────────────────────────
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0
werkzeug>=2.3.0

# ── ML / Vision ──────────────────────────────────────────────────────────────
tensorflow-cpu>=2.13.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# ── Model storage ─────────────────────────────────────────────────────────────
huggingface_hub>=0.20.0

# ── Audio ────────────────────────────────────────────────────────────────────
soundfile>=0.12.0
librosa>=0.10.0
